Bug fix: Crash when shutting down after installing an app
When you install an app, the `pumpkin_deploy_file()` function is executed. This function always loads the fonts from `BOOT.prc` every time it runs, which is a problem because this causes the system fonts to get re-loaded on a different thread. So, when the system shuts down the font memory is free’d from the wrong thread heap, causing a crash.

The fix is to only load the `BOOT.prc` fonts after deploying files from libos.c.